Add a closing div tag for .page-wrap
A close for .page-wrap was found to be missing when running HTML5 Validator on the output source. This commit adds it into footer.php. It was added above <footer> as it starts below the <header>. It was assumed that the intent is to wrap the page, sans header and footer.