Full-width menu support
Introduces support for the first menu format: full-width.
This menu format uses standard text links in a faux three column layout.
It should be used sparingly, because it creates L-shaped mouse tunnels. The standard drop down menu is a better UX.

This format only requires a css classname to be added to the top-level <li>, however
a custom walker was put into place to support more complex nav menu elements.