Fixes an issue with gallery caption cycling
When clicking next in the gallery paging, the caption would appear to be random, at the very least it wasn't the same as the photo. I tracked down the issue to line 30, which was triggering the caption to go to the nextSlide so very quickly it would become out of sync. Removing that line allows the data-attribute config of cycle2 to paginate things correctly.

This commit also fixes the issues raised by JSHint:
* Added a typed equality comparison with undefined
* Removed linkedIn from the things to be updated. That selector does not exist in the output. (e.g. LinkedIn is not currently displayed on the front-end
* Removed an isWidescreen() check because isWidescreen() does not exist and nor does expand_wide().
* Removed the global variable GMR_Gallery because the three methods it exposes do not exist.