replaces get_stylesheet_directory with proper template includes.
Replaces get_stylesheet_directory_uri() for the styles with get_template_directory. It seems counter-intuitive, but we want the parent theme styles by default. They can be overridden by a child theme. The functions are named for the location they point to, not necessarily for their intended use.

Replaces get_stylesheet_directory() for the profile page template with locate_template(). This allows a child theme to optionally override the parent theme's profile template, but does not make it mandatory.