Logo alignment fix
Fixes an issue where th logo wouldn't be vertically aligned with the menu. This was happening when the logo's aspect ratio became contained by the container width, versus it's height. The fix is to vertically center the image within the container *and* constrain by max-height and width.

This also provides a new variable $header-logo-padding so that a child theme may, more readily, change how much room is available to the logo. It's tweaked so that the WMGK logo looks nice in that spot, but more square logos may benefit from less padding.

The div that previously surrounded the logo and it's anchor was decided to be superfluous. Instead the anchor itself becomes the block level element tha constrains the image. This also made it easier to perform the vertical centering. All around a win.

Vertical centering likely won't work on IE8 due to the CSS rule used to make it happen. This will result in small alignment issues, which I think will be acceptable.