Keep the mobile menu around a bit longer to avoid cropped nav menus
The horizontal nav menu was being cropped at breakpoints between 768 (below which the mobile nav kicked in) and around 1020px even with a modest number of items. Although some additional room could be spared in the header by cramming items together (especially the nav menu against the logo) it wouldn't be enough to get the number of items we needed onto the line.

The solution is to keep the mobile menus around until a slightly larger breakpoint, 1024px. However, at 768 we put the search form and account menus up like we do at larger breakpoints, since there's room. Because of this, we also put the logo to the left, instead of centered (like it is on mobile). If it remained centered, it'd be a bit off balance.

The logo is designed to align vertically centered with the rest of the items which is good for generic logos. Some logos though, look better moved up or down a bit. For WMGK, I used absolute positioning to move the logo up a few pixels. YMMV