Style the keywords autocomplete
Modify the autocomplete init to append to the header search form and zero out the inline styles it automatically adds.

The style for the autocomplete form are derived in large part from the submenus, and where possible, the jquery ui classes were added to the list of selectors. But there is a new scss stylesheet for everything that still needs some help, layout especially.