Fixes an issue with full-width nav menus.
Full width nav menus are width:100%, but that means that if we have the top level <li>s, they will be 100% of the li. In reality, we want them to be 100% of the content max-width.

By putting the position:relative on header__main, the full-width sub menus will now be positioned relative to the header, not the LI — so they're full width. The downside, which is acceptable, is that the header content (logo, nav, search, account menu) are all on top of the overlay. So the focus effect isn't quite as pronounced.