aligned left and right media shouldn't take up the full width if there's lots of room.
This is ostensibly to solve an issue with the content needing to float around ads. By giving large floated images a max-width of 50%, they can sit next to the ad. If they are large images (i.e. a high max-width value) and floated, they will simply clear to the bottom of the ad.