Adds ooyala to fitvids and overrides some styles.
We will still need the ooyala responsive video plugin as this changes the shortcode itself. Unfortunately the output from the responsive plugin isn't quite as dynamic as fitvids.js — it doesn't dynamically calculate the aspect ratio. This is why we apply fitvids to ooyala.

However, fitvids applies a width of 100%, when we want a max-width so the content flows around the ads.