♻️ Update default store
I would prefer redux to manage the "current-menu-item" state, and we get
that info from the page load (or since this is a hybrid,a check after
componentDidMount).

Then managing this through thunks (since that is what we have
installed)or other side effects.

At the risk of doing a rewrite where we remove the checks completely and
instead rely on state, I am keeping this update as-is.

Please reach out with any questions or concerns.