🚨️ fix linter errors on this branch (+ comma-dangle)
Added `comma-dangle` as though that is not being used in `@10up/eslint`,
it is used extensively throughout the application. (This is turned on only for multi-line at the moment.)

Added `sort-keys` as this greatly enhances the human readability. However, due to the large nature of that particular lint update, have chosen to disable it for now. (No real gain at this point.)

We are extending from `eslint` so things like `max-len` has
a lot of updates as well. As well as the removing the empty line after a component is extended.

ref: https://eslint.org/docs/rules/max-len