♻️ Use Slugified URL over Timestamp
This was causing issues in IE, and in general.
There was a condition where UUID was not being passed consistently.
This requires a urlSlugified before dispatch to override what uuid
should be in the replace state.

This was a lot more diffiicult than it looks, haha.

Also, we are slugify'ing the hole url because adding the complexity of
just doing the pathname is beyond me at the moment.